+++ express/db/task.js
@@ -8,7 +8,7 @@ const taskSchema = new mongoose.Schema({
   },
   position: Number,
   taskComment: String,
-  createTime: String,
+  createTime: Number,
 }, {autoIndex: false, versionKey: false});
 
 const Task = mongoose.model('tasks', taskSchema);
