+++ express/helperFunctions.js
@@ -13,30 +13,8 @@ module.exports = {
         taskText: task.taskText,
         position: task.position,
         taskComment: task.taskComment,
+        createTime: task.createTime,
       }
     });
   },
-  getCurrentDateAsString() {
-    const date = new Date();
-    const months = [
-      'January',
-      'February',
-      'March',
-      'April',
-      'May',
-      'June',
-      'July',
-      'August',
-      'September',
-      'October',
-      'November',
-      'December',
-    ]
-    const month = months[date.getMonth()];
-    const day = date.getDate();
-    const year = date.getFullYear();
-    const hours = date.getHours();
-    const min = date.getMinutes();
-    return `${month} ${day}, ${year}, ${hours}:${min}`;
-  }
 };
