+++ src/types/types.ts
@@ -24,4 +24,5 @@ export type DeleteWarningdProps = {
 export type EditTaskDataType = {
   taskText: string;
   taskComment: string;
+  createTime?: number;
 };
